+++ docs/search/all_6.js
@@ -2,6 +2,8 @@ var searchData=
 [
   ['encryptcookies',['EncryptCookies',['../classBristolSU_1_1Playground_1_1Http_1_1Middleware_1_1EncryptCookies.html',1,'BristolSU::Playground::Http::Middleware']]],
   ['encryptcookies_2ephp',['EncryptCookies.php',['../EncryptCookies_8php.html',1,'']]],
+  ['event',['Event',['../classBristolSU_1_1Playground_1_1Support_1_1Events_1_1Event.html',1,'BristolSU::Playground::Support::Events']]],
+  ['event_2ephp',['Event.php',['../Event_8php.html',1,'']]],
   ['eventserviceprovider',['EventServiceProvider',['../classBristolSU_1_1Playground_1_1Providers_1_1EventServiceProvider.html',1,'BristolSU::Playground::Providers']]],
   ['eventserviceprovider_2ephp',['EventServiceProvider.php',['../EventServiceProvider_8php.html',1,'']]]
 ];