+++ docs/search/classes_2.js
@@ -1,5 +1,6 @@
 var searchData=
 [
   ['encryptcookies',['EncryptCookies',['../classBristolSU_1_1Playground_1_1Http_1_1Middleware_1_1EncryptCookies.html',1,'BristolSU::Playground::Http::Middleware']]],
+  ['event',['Event',['../classBristolSU_1_1Playground_1_1Support_1_1Events_1_1Event.html',1,'BristolSU::Playground::Support::Events']]],
   ['eventserviceprovider',['EventServiceProvider',['../classBristolSU_1_1Playground_1_1Providers_1_1EventServiceProvider.html',1,'BristolSU::Playground::Providers']]]
 ];