+++ docs/search/files_a.js
@@ -1,5 +1,6 @@
 var searchData=
 [
+  ['saveeventindatabase_2ephp',['SaveEventInDatabase.php',['../SaveEventInDatabase_8php.html',1,'']]],
   ['serviceconnectioncontroller_2ephp',['ServiceConnectionController.php',['../ServiceConnectionController_8php.html',1,'']]],
   ['serviceconnectorcontroller_2ephp',['ServiceConnectorController.php',['../ServiceConnectorController_8php.html',1,'']]]
 ];