+++ docs/search/functions_c.js
@@ -1,5 +1,4 @@
 var searchData=
 [
-  ['validatecredentials',['validateCredentials',['../classBristolSU_1_1Playground_1_1Support_1_1Authentication_1_1UserProvider.html#ad90aa5863a5a2777a5a6f167fba61742',1,'BristolSU::Playground::Support::Authentication::UserProvider']]],
-  ['validator',['validator',['../classBristolSU_1_1Playground_1_1Http_1_1Controllers_1_1Auth_1_1RegisterController.html#a01ce92e4781509b9397978440ecffb2a',1,'BristolSU::Playground::Http::Controllers::Auth::RegisterController']]]
+  ['test',['test',['../classBristolSU_1_1Playground_1_1Http_1_1Controllers_1_1Api_1_1ConnectionController.html#aef3efdd43ce96a3e508578b7b414e7d2',1,'BristolSU\Playground\Http\Controllers\Api\ConnectionController\test()'],['../classBristolSU_1_1Playground_1_1Http_1_1Controllers_1_1Api_1_1ModuleInstanceCompletionConditionController.html#a9b216ab13046c11ff20442cbca136821',1,'BristolSU\Playground\Http\Controllers\Api\ModuleInstanceCompletionConditionController\test()']]]
 ];